refactor(core): split configuration concerns
Part of #1108.
